controller success

+++ e-commerce-backend/src/main/java/com/ozaslan/e_commerce_backend/model/Cart.java
@@ -25,7 +25,7 @@ public class Cart {
     private double totalPrice;
 
     @Column(name = "total_Item")
-    private double totalItem;
+    private int totalItem;
 
     private int totalDiscountedPrice;
 
@@ -34,7 +34,7 @@ public class Cart {
     public Cart() {
     }
 
-    public Cart(Long id, User user, Set<CartItem> cartItems, double totalPrice, int totalDiscountedPrice, int discounte, double totalItem) {
+    public Cart(Long id, User user, Set<CartItem> cartItems, double totalPrice, int totalDiscountedPrice, int discounte, int totalItem) {
         this.id = id;
         this.user = user;
         this.cartItems = cartItems;
@@ -76,11 +76,11 @@ public class Cart {
         this.totalPrice = totalPrice;
     }
 
-    public double getTotalItem() {
+    public int getTotalItem() {
         return totalItem;
     }
 
-    public void setTotalItem(double totalItem) {
+    public void setTotalItem(int totalItem) {
         this.totalItem = totalItem;
     }
